job_recommendation-main/requirements.txt
# Core web framework
Flask==3.0.0
Werkzeug==3.0.1
Flask-SQLAlchemy==3.1.1
Flask-Limiter==3.5.0
Flask-Cors==4.0.0

# Database
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Data processing and ML
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
nltk==3.8.1

# Web scraping and document processing
beautifulsoup4==4.12.2
requests==2.31.0
selenium==4.15.2
PyPDF2==3.0.1
docx2txt==0.8
python-docx==1.0.1

# Utilities
python-dotenv==1.0.0

# Deployment
gunicorn==21.2.0